ORC-84. Create a separate java tool module
