+++ java/core/src/test/org/apache/orc/TestColumnStatistics.java
@@ -37,8 +37,6 @@ import org.apache.hadoop.hive.serde2.io.DateWritable;
 import org.apache.hadoop.io.BytesWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.orc.impl.ColumnStatisticsImpl;
-import org.apache.orc.tools.FileDump;
-import org.apache.orc.tools.TestFileDump;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -204,161 +202,5 @@ public class TestColumnStatistics {
     fs.setWorkingDirectory(workDir);
     testFilePath = new Path("TestOrcFile." + testCaseName.getMethodName() + ".orc");
     fs.delete(testFilePath, false);
-  }
-
-  private static BytesWritable bytes(int... items) {
-    BytesWritable result = new BytesWritable();
-    result.setSize(items.length);
-    for (int i = 0; i < items.length; ++i) {
-      result.getBytes()[i] = (byte) items[i];
-    }
-    return result;
-  }
-
-  void appendRow(VectorizedRowBatch batch, BytesWritable bytes,
-                 String str) {
-    int row = batch.size++;
-    if (bytes == null) {
-      batch.cols[0].noNulls = false;
-      batch.cols[0].isNull[row] = true;
-    } else {
-      ((BytesColumnVector) batch.cols[0]).setVal(row, bytes.getBytes(),
-          0, bytes.getLength());
-    }
-    if (str == null) {
-      batch.cols[1].noNulls = false;
-      batch.cols[1].isNull[row] = true;
-    } else {
-      ((BytesColumnVector) batch.cols[1]).setVal(row, str.getBytes());
-    }
-  }
-
-  @Test
-  public void testHasNull() throws Exception {
-    TypeDescription schema =
-        TypeDescription.createStruct()
-        .addField("bytes1", TypeDescription.createBinary())
-        .addField("string1", TypeDescription.createString());
-    Writer writer = OrcFile.createWriter(testFilePath,
-        OrcFile.writerOptions(conf)
-            .setSchema(schema)
-            .rowIndexStride(1000)
-            .stripeSize(10000)
-            .bufferSize(10000));
-    VectorizedRowBatch batch = schema.createRowBatch(5000);
-    // STRIPE 1
-    // RG1
-    for(int i=0; i<1000; i++) {
-      appendRow(batch, bytes(1, 2, 3), "RG1");
-    }
-    writer.addRowBatch(batch);
-    batch.reset();
-    // RG2
-    for(int i=0; i<1000; i++) {
-      appendRow(batch, bytes(1, 2, 3), null);
-    }
-    writer.addRowBatch(batch);
-    batch.reset();
-    // RG3
-    for(int i=0; i<1000; i++) {
-      appendRow(batch, bytes(1, 2, 3), "RG3");
-    }
-    writer.addRowBatch(batch);
-    batch.reset();
-    // RG4
-    for (int i = 0; i < 1000; i++) {
-      appendRow(batch, bytes(1,2,3), null);
-    }
-    writer.addRowBatch(batch);
-    batch.reset();
-    // RG5
-    for(int i=0; i<1000; i++) {
-      appendRow(batch, bytes(1, 2, 3), null);
-    }
-    writer.addRowBatch(batch);
-    batch.reset();
-    // STRIPE 2
-    for (int i = 0; i < 5000; i++) {
-      appendRow(batch, bytes(1,2,3), null);
-    }
-    writer.addRowBatch(batch);
-    batch.reset();
-    // STRIPE 3
-    for (int i = 0; i < 5000; i++) {
-      appendRow(batch, bytes(1,2,3), "STRIPE-3");
-    }
-    writer.addRowBatch(batch);
-    batch.reset();
-    // STRIPE 4
-    for (int i = 0; i < 5000; i++) {
-      appendRow(batch, bytes(1,2,3), null);
-    }
-    writer.addRowBatch(batch);
-    batch.reset();
-    writer.close();
-    Reader reader = OrcFile.createReader(testFilePath,
-        OrcFile.readerOptions(conf).filesystem(fs));
-
-    // check the file level stats
-    ColumnStatistics[] stats = reader.getStatistics();
-    assertEquals(20000, stats[0].getNumberOfValues());
-    assertEquals(20000, stats[1].getNumberOfValues());
-    assertEquals(7000, stats[2].getNumberOfValues());
-    assertEquals(false, stats[0].hasNull());
-    assertEquals(false, stats[1].hasNull());
-    assertEquals(true, stats[2].hasNull());
-
-    // check the stripe level stats
-    List<StripeStatistics> stripeStats = reader.getStripeStatistics();
-    // stripe 1 stats
-    StripeStatistics ss1 = stripeStats.get(0);
-    ColumnStatistics ss1_cs1 = ss1.getColumnStatistics()[0];
-    ColumnStatistics ss1_cs2 = ss1.getColumnStatistics()[1];
-    ColumnStatistics ss1_cs3 = ss1.getColumnStatistics()[2];
-    assertEquals(false, ss1_cs1.hasNull());
-    assertEquals(false, ss1_cs2.hasNull());
-    assertEquals(true, ss1_cs3.hasNull());
-
-    // stripe 2 stats
-    StripeStatistics ss2 = stripeStats.get(1);
-    ColumnStatistics ss2_cs1 = ss2.getColumnStatistics()[0];
-    ColumnStatistics ss2_cs2 = ss2.getColumnStatistics()[1];
-    ColumnStatistics ss2_cs3 = ss2.getColumnStatistics()[2];
-    assertEquals(false, ss2_cs1.hasNull());
-    assertEquals(false, ss2_cs2.hasNull());
-    assertEquals(true, ss2_cs3.hasNull());
-
-    // stripe 3 stats
-    StripeStatistics ss3 = stripeStats.get(2);
-    ColumnStatistics ss3_cs1 = ss3.getColumnStatistics()[0];
-    ColumnStatistics ss3_cs2 = ss3.getColumnStatistics()[1];
-    ColumnStatistics ss3_cs3 = ss3.getColumnStatistics()[2];
-    assertEquals(false, ss3_cs1.hasNull());
-    assertEquals(false, ss3_cs2.hasNull());
-    assertEquals(false, ss3_cs3.hasNull());
-
-    // stripe 4 stats
-    StripeStatistics ss4 = stripeStats.get(3);
-    ColumnStatistics ss4_cs1 = ss4.getColumnStatistics()[0];
-    ColumnStatistics ss4_cs2 = ss4.getColumnStatistics()[1];
-    ColumnStatistics ss4_cs3 = ss4.getColumnStatistics()[2];
-    assertEquals(false, ss4_cs1.hasNull());
-    assertEquals(false, ss4_cs2.hasNull());
-    assertEquals(true, ss4_cs3.hasNull());
-
-    // Test file dump
-    PrintStream origOut = System.out;
-    String outputFilename = "orc-file-has-null.out";
-    FileOutputStream myOut = new FileOutputStream(workDir + File.separator + outputFilename);
-
-    // replace stdout and run command
-    System.setOut(new PrintStream(myOut));
-    FileDump.main(new String[]{testFilePath.toString(), "--rowindex=2"});
-    System.out.flush();
-    System.setOut(origOut);
-    // If called with an expression evaluating to false, the test will halt
-    // and be ignored.
-    assumeTrue(!System.getProperty("os.name").startsWith("Windows"));
-    TestFileDump.checkOutput(outputFilename, workDir + File.separator + outputFilename);
   }
 }
