+++ java/storage-api/pom.xml
@@ -20,6 +20,7 @@
     <groupId>org.apache</groupId>
     <artifactId>apache</artifactId>
     <version>18</version>
+    <relativePath></relativePath>
   </parent>
 
   <groupId>org.apache.hive</groupId>
