+++ java/tools/src/java/org/apache/orc/tools/FileDump.java
@@ -18,8 +18,6 @@
 package org.apache.orc.tools;
 
 import java.io.IOException;
-import java.io.OutputStreamWriter;
-import java.io.PrintStream;
 import java.text.DecimalFormat;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -39,24 +37,11 @@ import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.fs.PathFilter;
 import org.apache.hadoop.hdfs.DistributedFileSystem;
-import org.apache.hadoop.hive.ql.exec.vector.BytesColumnVector;
-import org.apache.hadoop.hive.ql.exec.vector.ColumnVector;
-import org.apache.hadoop.hive.ql.exec.vector.DecimalColumnVector;
-import org.apache.hadoop.hive.ql.exec.vector.DoubleColumnVector;
-import org.apache.hadoop.hive.ql.exec.vector.ListColumnVector;
-import org.apache.hadoop.hive.ql.exec.vector.LongColumnVector;
-import org.apache.hadoop.hive.ql.exec.vector.MapColumnVector;
-import org.apache.hadoop.hive.ql.exec.vector.StructColumnVector;
-import org.apache.hadoop.hive.ql.exec.vector.TimestampColumnVector;
-import org.apache.hadoop.hive.ql.exec.vector.UnionColumnVector;
-import org.apache.hadoop.hive.ql.exec.vector.VectorizedRowBatch;
-import org.apache.hadoop.hive.serde2.io.DateWritable;
 import org.apache.orc.BloomFilterIO;
 import org.apache.orc.ColumnStatistics;
 import org.apache.orc.CompressionKind;
 import org.apache.orc.OrcFile;
 import org.apache.orc.Reader;
-import org.apache.orc.RecordReader;
 import org.apache.orc.TypeDescription;
 import org.apache.orc.Writer;
 import org.apache.orc.impl.AcidStats;
@@ -67,8 +52,6 @@ import org.apache.orc.OrcProto;
 import org.apache.orc.StripeInformation;
 import org.apache.orc.StripeStatistics;
 import org.apache.orc.impl.RecordReaderImpl;
-import org.codehaus.jettison.json.JSONException;
-import org.codehaus.jettison.json.JSONWriter;
 
 /**
  * A tool for printing out the file structure of ORC files.
@@ -90,9 +73,7 @@ public final class FileDump {
   private FileDump() {
   }
 
-  public static void main(String[] args) throws Exception {
-    Configuration conf = new Configuration();
-
+  public static void main(Configuration conf, String[] args) throws Exception {
     List<Integer> rowIndexCols = new ArrayList<Integer>(0);
     Options opts = createOptions();
     CommandLine cli = new GnuParser().parse(opts, args);
@@ -140,7 +121,7 @@ public final class FileDump {
     }
 
     if (dumpData) {
-      printData(filesInPath, conf);
+      PrintData.main(conf, filesInPath.toArray(new String[filesInPath.size()]));
     } else if (recover && skipDump) {
       recoverFiles(filesInPath, conf, backupPath);
     } else {
@@ -151,6 +132,11 @@ public final class FileDump {
         printMetaData(filesInPath, conf, rowIndexCols, printTimeZone, recover, backupPath);
       }
     }
+  }
+
+  public static void main(String[] args) throws Exception {
+    Configuration conf = new Configuration();
+    main(conf, args);
   }
 
   /**
@@ -274,25 +260,6 @@ public final class FileDump {
     }
 
     return filesInPath;
-  }
-
-  private static void printData(List<String> files,
-      Configuration conf) throws IOException,
-      JSONException {
-    for (String file : files) {
-      try {
-        Path path = new Path(file);
-        Reader reader = getReader(path, conf, new ArrayList<String>());
-        if (reader == null) {
-          continue;
-        }
-        printJsonData(reader);
-        System.out.println(SEPARATOR);
-      } catch (Exception e) {
-        System.err.println("Unable to dump data for file: " + file);
-        continue;
-      }
-    }
   }
 
   private static void printMetaData(List<String> files, Configuration conf,
@@ -782,167 +749,4 @@ public final class FileDump {
     return result;
   }
 
-  private static void printMap(JSONWriter writer,
-                               MapColumnVector vector,
-                               TypeDescription schema,
-                               int row) throws JSONException {
-    writer.array();
-    TypeDescription keyType = schema.getChildren().get(0);
-    TypeDescription valueType = schema.getChildren().get(1);
-    int offset = (int) vector.offsets[row];
-    for (int i = 0; i < vector.lengths[row]; ++i) {
-      writer.object();
-      writer.key("_key");
-      printValue(writer, vector.keys, keyType, offset + i);
-      writer.key("_value");
-      printValue(writer, vector.values, valueType, offset + i);
-      writer.endObject();
-    }
-    writer.endArray();
-  }
-
-  private static void printList(JSONWriter writer,
-                                ListColumnVector vector,
-                                TypeDescription schema,
-                                int row) throws JSONException {
-    writer.array();
-    int offset = (int) vector.offsets[row];
-    TypeDescription childType = schema.getChildren().get(0);
-    for (int i = 0; i < vector.lengths[row]; ++i) {
-      printValue(writer, vector.child, childType, offset + i);
-    }
-    writer.endArray();
-  }
-
-  private static void printUnion(JSONWriter writer,
-                                 UnionColumnVector vector,
-                                 TypeDescription schema,
-                                 int row) throws JSONException {
-    int tag = vector.tags[row];
-    printValue(writer, vector.fields[tag], schema.getChildren().get(tag), row);
-  }
-
-  static void printStruct(JSONWriter writer,
-                          StructColumnVector batch,
-                          TypeDescription schema,
-                          int row) throws JSONException {
-    writer.object();
-    List<String> fieldNames = schema.getFieldNames();
-    List<TypeDescription> fieldTypes = schema.getChildren();
-    for (int i = 0; i < fieldTypes.size(); ++i) {
-      writer.key(fieldNames.get(i));
-      printValue(writer, batch.fields[i], fieldTypes.get(i), row);
-    }
-    writer.endObject();
-  }
-
-  static void printBinary(JSONWriter writer, BytesColumnVector vector,
-                          int row) throws JSONException {
-    writer.array();
-    int offset = vector.start[row];
-    for(int i=0; i < vector.length[row]; ++i) {
-      writer.value(0xff & (int) vector.vector[row][offset + i]);
-    }
-    writer.endArray();
-  }
-  static void printValue(JSONWriter writer, ColumnVector vector,
-                         TypeDescription schema, int row) throws JSONException {
-    if (vector.isRepeating) {
-      row = 0;
-    }
-    if (vector.noNulls || !vector.isNull[row]) {
-      switch (schema.getCategory()) {
-        case BOOLEAN:
-          writer.value(((LongColumnVector) vector).vector[row] != 0);
-          break;
-        case BYTE:
-        case SHORT:
-        case INT:
-        case LONG:
-          writer.value(((LongColumnVector) vector).vector[row]);
-          break;
-        case FLOAT:
-        case DOUBLE:
-          writer.value(((DoubleColumnVector) vector).vector[row]);
-          break;
-        case STRING:
-        case CHAR:
-        case VARCHAR:
-          writer.value(((BytesColumnVector) vector).toString(row));
-          break;
-        case BINARY:
-          printBinary(writer, (BytesColumnVector) vector, row);
-          break;
-        case DECIMAL:
-          writer.value(((DecimalColumnVector) vector).vector[row].toString());
-          break;
-        case DATE:
-          writer.value(new DateWritable(
-              (int) ((LongColumnVector) vector).vector[row]).toString());
-          break;
-        case TIMESTAMP:
-          writer.value(((TimestampColumnVector) vector)
-              .asScratchTimestamp(row).toString());
-          break;
-        case LIST:
-          printList(writer, (ListColumnVector) vector, schema, row);
-          break;
-        case MAP:
-          printMap(writer, (MapColumnVector) vector, schema, row);
-          break;
-        case STRUCT:
-          printStruct(writer, (StructColumnVector) vector, schema, row);
-          break;
-        case UNION:
-          printUnion(writer, (UnionColumnVector) vector, schema, row);
-          break;
-        default:
-          throw new IllegalArgumentException("Unknown type " +
-              schema.toString());
-      }
-    } else {
-      writer.value(null);
-    }
-  }
-
-  static void printRow(JSONWriter writer,
-                       VectorizedRowBatch batch,
-                       TypeDescription schema,
-                       int row) throws JSONException {
-    if (schema.getCategory() == TypeDescription.Category.STRUCT) {
-      List<TypeDescription> fieldTypes = schema.getChildren();
-      List<String> fieldNames = schema.getFieldNames();
-      writer.object();
-      for (int c = 0; c < batch.cols.length; ++c) {
-        writer.key(fieldNames.get(c));
-        printValue(writer, batch.cols[c], fieldTypes.get(c), row);
-      }
-      writer.endObject();
-    } else {
-      printValue(writer, batch.cols[0], schema, row);
-    }
-  }
-
-  static void printJsonData(final Reader reader) throws IOException, JSONException {
-    PrintStream printStream = System.out;
-    OutputStreamWriter out = new OutputStreamWriter(printStream, "UTF-8");
-    RecordReader rows = reader.rows();
-    try {
-      TypeDescription schema = reader.getSchema();
-      VectorizedRowBatch batch = schema.createRowBatch();
-      while (rows.nextBatch(batch)) {
-        for(int r=0; r < batch.size; ++r) {
-          JSONWriter writer = new JSONWriter(out);
-          printRow(writer, batch, schema, r);
-          out.write("\n");
-          out.flush();
-          if (printStream.checkError()) {
-            throw new IOException("Error encountered when writing to stdout.");
-          }
-        }
-      }
-    } finally {
-      rows.close();
-    }
-  }
 }
